# Enhenced String report and stacked bar visulization.