… figures/test.png
… figures/test_vit.png